Separate contract and business APIs in contract
