+++ smart-contract/hyperledger-fabric/v2/java/src/main/java/hu/bme/mit/ftsrg/chaincode/tpcc/api/NotFoundException.java
@@ -1,0 +1,8 @@
+package hu.bme.mit.ftsrg.chaincode.tpcc.api;
+
+public final class NotFoundException extends Exception {
+
+  NotFoundException(String message) {
+    super(message);
+  }
+}
