+++ smart-contract/hyperledger-fabric/v2/java/src/main/java/hu/bme/mit/ftsrg/chaincode/tpcc/data/input/OrderStatusInput.java
@@ -2,8 +2,6 @@
 
 package hu.bme.mit.ftsrg.chaincode.tpcc.data.input;
 
-import hu.bme.mit.ftsrg.chaincode.tpcc.TPCC;
-import hu.bme.mit.ftsrg.chaincode.tpcc.middleware.TPCCContext;
 import lombok.EqualsAndHashCode;
 import org.hyperledger.fabric.contract.annotation.DataType;
 import org.hyperledger.fabric.contract.annotation.Property;
