+++ smart-contract/hyperledger-fabric/v2/java/src/main/java/hu/bme/mit/ftsrg/chaincode/tpcc/data/output/OrderStatusOutput.java
@@ -2,11 +2,9 @@
 
 package hu.bme.mit.ftsrg.chaincode.tpcc.data.output;
 
-import hu.bme.mit.ftsrg.chaincode.tpcc.TPCC;
 import hu.bme.mit.ftsrg.chaincode.tpcc.data.entity.Customer;
 import hu.bme.mit.ftsrg.chaincode.tpcc.data.entity.Order;
 import hu.bme.mit.ftsrg.chaincode.tpcc.data.extra.OrderLineData;
-import hu.bme.mit.ftsrg.chaincode.tpcc.middleware.TPCCContext;
 import java.util.List;
 import lombok.EqualsAndHashCode;
 import org.hyperledger.fabric.contract.annotation.DataType;
